pyproject.toml
[project]
name = "on-prem-embedding-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch",
    "transformers",
    "fastapi",
    "uvicorn",
    "optimum[onnxruntime-gpu]",
    "onnx-graphsurgeon",
    "polygraphy",
]
